+++ src/data/userGroups.ts
@@ -4,11 +4,11 @@ import { lang } from "../services/languageService";
 const UserGroups = new Map<UserGroupEnum, any>();
 UserGroups.set(UserGroupEnum.blindPeople, {
   name: lang.userProfileVisImpairments,
-  icon: "\\images\\eye.svg",
+  icon: "/images/eye.svg",
 });
 UserGroups.set(UserGroupEnum.noImpairments, {
   name: lang.userProfileNoSpecialNeeds,
-  icon: "\\images\\nothing.svg",
+  icon: "/images/nothing.svg",
 });
 UserGroups.set(UserGroupEnum.wheelchairUsers, {
   name: lang.userProfileWheelchair,